requirements.txt
PyMuPDF
PyPDF2
pdfminer.six
